requirements.txt
python-pptx
Pillow
pymupdf
pillow-heif
